Metadata as a constexpr POD in read-only memory, IObjectDictionnary removed
Metadata is now a plain struct {access, dataType, size, defaultValue, low,
high} built at compile time by makeMetadata(), makeStringMetadata() and
makeDomainMetadata() (src/od/metadata.hpp). The generated dictionary emits
one constexpr constant per default value and limit, and a const table of
Metadata values: no vtable, no std::string static, no dynamic
initialisation, the table lands in .const on the target. getMetadata()
returns a const pointer and getSize() reads the size from the metadata, so
a VISIBLE_STRING reports its capacity.

The generated od.hpp defines OD_<TYPE>_COUNT for each typed data table and
only declares the tables that exist; od_common.cpp defines the matching
accessors under the same guard, so a dictionary without a given type no
longer produces zero-size arrays.

Macros of od_common.hpp used only by od_common.cpp become constexpr
constants there; the unused OD_GRANULARITY, OD_ENTRY_SIZE_MAX,
OBJECT_INDEX_COUNT and X1A00_* are dropped. The -Wconversion warning of
EMCY::raiseError is fixed on the way.

Contract test extended with size, limits, string and domain metadata. The
read-only SDO test targets 0x1000 instead of patching the metadata.

+++ platform/linux/FilePersistence.cpp
@@ -24,7 +24,7 @@ bool FilePersistence::saveGroup(uint8_t parameterGroup) {
     std::ofstream file(groupPath(parameterGroup),
                        std::ios::out | std::ios::binary | std::ios::trunc);
     if (!file) return false;
-    for (int32_t id = 0; id < static_cast<int32_t>(OD_LENGTH); id++) {
+    for (int32_t id = 0; id < OD_LENGTH; id++) {
         if (!inParameterGroup(CANopenOD::objectIndexTable[id].first,
                               parameterGroup))
             continue;
@@ -42,7 +42,7 @@ bool FilePersistence::loadGroup(uint8_t parameterGroup) {
     std::ifstream file(groupPath(parameterGroup),
                        std::ios::in | std::ios::binary);
     if (!file) return false;
-    for (int32_t id = 0; id < static_cast<int32_t>(OD_LENGTH); id++) {
+    for (int32_t id = 0; id < OD_LENGTH; id++) {
         if (!inParameterGroup(CANopenOD::objectIndexTable[id].first,
                               parameterGroup))
             continue;
